rimossa classe AnswerRepository rimossa classe AnswerRepository e delegato il compito alla classe QuestionRepository, adattato i vari pezzi di codice di conseguenza
